+++ src/components/Chat/Chat.tsx
@@ -13,7 +13,7 @@ import {
   ShorthandRenderFunction,
 } from '../../../types/utils'
 import { Accessibility, AccessibilityActionHandlers } from '../../lib/accessibility/interfaces'
-import ChatBehavior from '../../lib/accessibility/Behaviors/Chat/ChatBehavior'
+import { chatBehavior } from '../../lib/accessibility'
 
 export interface IChatProps {
   accessibility?: Accessibility
@@ -67,7 +67,7 @@ class Chat extends UIComponent<Extendable<IChatProps>, any> {
     variables: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   }
 
-  static defaultProps = { accessibility: ChatBehavior as Accessibility, as: 'ul' }
+  static defaultProps = { accessibility: chatBehavior as Accessibility, as: 'ul' }
 
   static Item = ChatItem
   static Message = ChatMessage