+++ src/lib/accessibility/Behaviors/Button/buttonGroupBehavior.ts
@@ -5,7 +5,7 @@ import { Accessibility, FocusZoneMode } from '../../interfaces'
  * Adds role 'presentation' to 'root' component's part.
  * Wraps component in FocusZone allowing arrow key navigation through the children of the component.
  */
-const ButtonGroupBehavior: Accessibility = {
+const buttonGroupBehavior: Accessibility = (props: any) => ({
   attributes: {
     root: {
       role: 'presentation',
@@ -18,6 +18,6 @@ const ButtonGroupBehavior: Accessibility = {
       preventDefaultWhenHandled: true,
     },
   },
-}
+})
 
-export default ButtonGroupBehavior
+export default buttonGroupBehavior