+++ src/lib/accessibility/Behaviors/Menu/menuBehavior.ts
@@ -6,7 +6,7 @@ import { Accessibility, FocusZoneMode } from '../../interfaces'
  * The 'menu' role is used to identify an element that creates a list of common actions or functions that a user can invoke.
  */
 
-const MenuBehavior: Accessibility = {
+const menuBehavior: Accessibility = (props: any) => ({
   attributes: {
     root: {
       role: 'menu',
@@ -19,6 +19,6 @@ const MenuBehavior: Accessibility = {
       preventDefaultWhenHandled: true,
     },
   },
-}
+})
 
-export default MenuBehavior
+export default menuBehavior